# Improve performance for fetching authorized entries