--- a/pkg/server/api/api.go
+++ b/pkg/server/api/api.go
@@ -18,7 +18,7 @@ type AuthorizedEntryFetcher interface {
 	LookupAuthorizedEntries(ctx context.Context, id spiffeid.ID, entryIDs map[string]struct{}) (map[string]ReadOnlyEntry, error)
 	// FetchAuthorizedEntries fetches the entries that the specified
 	// SPIFFE ID is authorized for
-	FetchAuthorizedEntries(ctx context.Context, id spiffeid.ID) ([]*types.Entry, error)
+	FetchAuthorizedEntries(ctx context.Context, id spiffeid.ID) ([]ReadOnlyEntry, error)
 }
 
 // AttestedNodeToProto converts an agent from the given *common.AttestedNode with