--- a/pkg/server/api/entry/v1/service.go
+++ b/pkg/server/api/entry/v1/service.go
@@ -390,6 +390,7 @@ func (s *Service) GetAuthorizedEntries(ctx context.Context, req *entryv1.GetAuth
 	if err != nil {
 		return nil, err
 	}
+
 	for i, entry := range entries {
 		applyMask(entry, req.OutputMask)
 		entries[i] = entry