@sorindumitru 2025-05-20
One thing I realized was that it's somewhat possible to switch the trust domain of a deployment. I've added some changes to filter the entries with the wrong trust domain from the cache and a test for that too.